[env] Added dtype to observation_space definition

+++ torille/envs/gym_env.py
@@ -63,7 +63,7 @@ class ToriEnv(gym.Env):
             )
 
         # For both players, position of all joints
-        self.observation_space = spaces.Box(low=-30, high=30, shape=(2,torille.ToribashConstants.NUM_LIMBS*3))
+        self.observation_space = spaces.Box(low=-30, high=30, dtype=np.float32, shape=(2,torille.ToribashConstants.NUM_LIMBS*3))
 
     def _preprocess_observation(self, state):
         """
